Make worker identity verification reviewable in one narrow path
The PR review was blocked not by the runtime/scaling intent, but by
verification packaging: the changed identity contract was only provable
through broader suites and ad-hoc review-worktree commands.

This adds a dedicated worker-runtime-identity test surface and a
single-purpose npm script so startup/scaling identity preservation can
be re-run cleanly from a fresh review checkout without depending on the
broader runtime/scaling suites for proof.

Constraint: Keep the runtime/scaling behavior unchanged while tightening only the verification story
Rejected: Re-explain the existing broad test commands in PR comments | still leaves the contract proof buried in large suites
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep future worker identity changes covered by the dedicated narrow verification path, not only the broad runtime/scaling suites
Tested: npm run test:team:worker-runtime-identity
Not-tested: Full repository test suite
