+++ src/team/__tests__/scaling.test.ts
@@ -570,94 +570,6 @@ exit 0
     }
   });
 
-  it('scaleUp preserves low-complexity assigned roles as outer runtime identities', async () => {
-    const cwd = await mkdtemp(join(tmpdir(), 'omx-scale-up-low-role-'));
-    const fakeBinDir = await mkdtemp(join(tmpdir(), 'omx-scale-up-low-role-bin-'));
-    const tmuxStubPath = join(fakeBinDir, 'tmux');
-    const previousPath = process.env.PATH;
-
-    try {
-      await writeFile(
-        tmuxStubPath,
-        `#!/bin/sh
-set -eu
-case "\${1:-}" in
-  -V)
-    echo "tmux 3.2a"
-    ;;
-  split-window)
-    echo "%31"
-    ;;
-  list-panes)
-    echo "42424"
-    ;;
-  capture-pane)
-    echo ""
-    ;;
-esac
-exit 0
-`,
-      );
-      await chmod(tmuxStubPath, 0o755);
-      process.env.PATH = `${fakeBinDir}:${previousPath ?? ''}`;
-
-      await mkdir(join(cwd, '.codex', 'prompts'), { recursive: true });
-      await writeFile(join(cwd, '.codex', 'prompts', 'explore.md'), '<identity>You are Explorer.</identity>');
-      await writeFile(join(cwd, '.codex', 'prompts', 'sisyphus-lite.md'), '<identity>You are Sisyphus-lite.</identity>');
-      await mkdir(join(cwd, '.omx', 'state', 'team', 'low-role-scale'), { recursive: true });
-      await writeFile(join(cwd, '.omx', 'state', 'team', 'low-role-scale', 'worker-agents.md'), '# Base worker instructions\n');
-
-      await initTeamState('low-role-scale', 'task', 'executor', 1, cwd, undefined, process.env, {
-        workspace_mode: 'single',
-        leader_cwd: cwd,
-        team_state_root: join(cwd, '.omx', 'state'),
-      });
-      await createTask('low-role-scale', {
-        subject: 'existing task',
-        description: 'already persisted',
-        status: 'pending',
-        owner: 'worker-1',
-      }, cwd);
-
-      const config = await readTeamConfig('low-role-scale', cwd);
-      assert.ok(config);
-      if (!config) return;
-      config.tmux_session = 'omx-team-low-role-scale';
-      config.leader_pane_id = '%11';
-      config.workers[0]!.pane_id = '%21';
-      await saveTeamConfig(config, cwd);
-
-      const manifestPath = join(cwd, '.omx', 'state', 'team', 'low-role-scale', 'manifest.v2.json');
-      const manifest = JSON.parse(await readFile(manifestPath, 'utf-8')) as { policy?: Record<string, unknown> };
-      manifest.policy = {
-        ...(manifest.policy ?? {}),
-        dispatch_mode: 'transport_direct',
-      };
-      await writeFile(manifestPath, JSON.stringify(manifest, null, 2));
-
-      const result = await scaleUp(
-        'low-role-scale',
-        1,
-        'executor',
-        [{ subject: 'map files', description: 'map files', owner: 'worker-2', role: 'explore' }],
-        cwd,
-        { OMX_TEAM_SCALING_ENABLED: '1', OMX_TEAM_SKIP_READY_WAIT: '1' },
-      );
-      assert.equal(result.ok, true);
-      if (!result.ok) return;
-
-      const workerAgents = await readFile(join(cwd, '.omx', 'state', 'team', 'low-role-scale', 'workers', 'worker-2', 'AGENTS.md'), 'utf-8');
-      assert.match(workerAgents, /You are operating as the \*\*explore\*\* role/);
-      assert.match(workerAgents, /You are Explorer\./);
-      assert.doesNotMatch(workerAgents, /Sisyphus-lite/);
-    } finally {
-      if (typeof previousPath === 'string') process.env.PATH = previousPath;
-      else delete process.env.PATH;
-      await rm(cwd, { recursive: true, force: true });
-      await rm(fakeBinDir, { recursive: true, force: true });
-    }
-  });
-
   it('does not apply mini guidance during scale-up when the final worker model is gpt-5.4', async () => {
     const cwd = await mkdtemp(join(tmpdir(), 'omx-scale-up-frontier-role-'));
     const fakeBinDir = await mkdtemp(join(tmpdir(), 'omx-scale-up-frontier-role-bin-'));
